# docs/architecture-pack/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Revizr API
  version: "1.0.0"
  description: |
    Revizr backend API — exam revision platform for UK students (KS3, GCSE, A-Level, 11+).
    All endpoints require Bearer JWT authentication unless noted as public.
    Base URL: https://api.revizr.co.uk/v1
  contact:
    name: Revizr Engineering
    email: engineering@revizr.co.uk
  license:
    name: Proprietary
servers:
  - url: https://api.revizr.co.uk/v1
    description: Production (eu-west-2)
  - url: https://api.staging.revizr.co.uk/v1
    description: Staging (eu-west-2)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by POST /auth/login or POST /auth/refresh.
        Expires in 15 minutes. Pass in Authorization header as "Bearer <token>".
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email is required"
        details:
          type: object
          example: { "field": "email", "constraint": "format" }
...
    NotificationPreferences:
      type: object
      required: [email_session_reminders, email_progress_reports, push_enabled]
      properties:
        email_session_reminders:
          type: boolean
          example: true
        email_progress_reports:
          type: boolean
          example: true
        push_enabled:
          type: boolean
          example: false
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired access token"
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "VALIDATION_ERROR"
            message: "Request body failed validation"
...
paths:
...
  /notifications/preferences:
    get:
      tags: [notifications]
      operationId: getNotificationPreferences
      summary: Get current user's notification preferences
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [notifications]
      operationId: updateNotificationPreferences
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_session_reminders:
                  type: boolean
                  example: false
                email_progress_reports:
                  type: boolean
                  example: true
                push_enabled:
                  type: boolean
                  example: true
                push_subscription:
                  type: object
                  description: Web Push API subscription object; required when push_enabled=true
                  properties:
                    endpoint:
                      type: string
                      format: uri
                      example: "https://fcm.googleapis.com/fcm/send/..."
                    keys:
                      type: object
                      properties:
                        p256dh:
                          type: string
                          example: "BNcRdreALRFXTkOOUHK1EtK2wtaz5Ry4YfYCA_0QTpQtUbVlTiE..."
                        auth:
                          type: string
                          example: "tBHItJI5svbpez7KI4CCXg=="
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...